feat: add responsive design and mobile burger menu
