+++ src/modules/shared/components/Layout/Header/Header.module.scss
@@ -1,30 +1,49 @@
 @import '../../../../../styles/variables';
 
 .header {
+  position: fixed;
   width: 100%;
   background-color: $background;
   border-bottom: 1px solid $surface-2;
-  transition: background-color 0.3s ease border-color 0.3s ease;
+  box-shadow: $shadow;
+  transition:
+    background-color 0.3s ease,
+    border-color 0.3s ease,
+    box-shadow 0.3s ease;
+  z-index: 100;
 }
 
 .container {
-  max-width: 1440px;
-  height: 64px;
   display: flex;
-  flex-direction: row;
   align-items: center;
+  justify-content: space-between;
+  max-width: 1440px;
+  width: 100%;
+  height: 48px;
+  padding: 0 16px;
   margin: 0 auto;
+
+  @media (min-width: 1024px) {
+    height: 64px;
+    padding: 0 32px;
+    flex-wrap: nowrap;
+  }
 }
 
 .headerContent {
-  padding: 0 24px;
+  display: flex;
+  align-items: center;
+  height: 100%;
 }
 
 .logo {
-  width: 80px;
-  height: 28px;
+  width: 64px;
   display: block;
   transition: opacity 0.3s ease;
+
+  @media (min-width: 1024px) {
+    width: 80px;
+  }
 
   &:hover,
   &:focus-visible {
@@ -33,92 +52,161 @@
 }
 
 .nav {
-  margin-right: 24px;
-  margin-left: 24px;
-  height: 100%;
+  position: fixed;
+  top: 48px;
+  right: -100%;
+  width: 100%;
+  height: calc(50vh - 24px);
+  background-color: $background;
+  padding: 24px;
+  box-sizing: border-box;
+  z-index: 99;
+  overflow-y: auto;
+  transition: right 0.3s ease;
+
+  &.navOpen {
+    right: 0;
+  }
+
+  @media (min-width: 768px) {
+    position: static;
+    width: auto;
+    height: 100%;
+    background-color: transparent;
+    padding: 0;
+    margin: 0 24px;
+    overflow: visible;
+  }
 }
 
 .list {
   display: flex;
-  height: 100%;
-  gap: 64px;
+  flex-direction: column;
+  gap: 24px;
   list-style: none;
   margin: 0;
   padding: 0;
+
+  @media (min-width: 768px) {
+    flex-direction: row;
+    height: 100%;
+    gap: 32px;
+  }
+
+  @media (min-width: 1024px) {
+    gap: 64px;
+  }
 }
 
 .link {
   display: flex;
   align-items: center;
+  width: fit-content;
   height: 100%;
+  padding-bottom: 8px;
   text-decoration: none;
   text-transform: uppercase;
-
-  @include uppercase;
-
   color: $text-secondary;
   border-bottom: 3px solid transparent;
   transition:
     color 0.3s ease,
     border-color 0.3s ease;
 
+  @include uppercase;
+
   &:hover {
     color: $text-main;
   }
+
+  @media (min-width: 768px) {
+    padding: 0;
+  }
 }
 
-.link.isActive {
+.isActive {
   color: $text-main;
   border-bottom-color: $accent;
+}
+
+.navActions {
+  position: fixed;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  top: calc(48px + (50vh - 24px));
+  right: -100%;
+  width: 100%;
+  height: calc(50vh - 24px);
+  background-color: $background;
+  border-top: 1px solid $surface-2;
+  box-sizing: border-box;
+  padding: 24px;
+  transition: right 0.3s ease;
+  z-index: 99;
+
+  &.navOpen {
+    right: 0;
+  }
+
+  @media (min-width: 768px) {
+    position: static;
+    width: auto;
+    height: 100%;
+    background-color: transparent;
+    border-top: none;
+    padding: 0;
+    margin-left: auto;
+  }
+}
+
+.togglerWrapper {
+  @media (min-width: 768px) {
+    display: flex;
+    align-items: center;
+    height: 100%;
+    margin-left: auto;
+    border: none;
+  }
 }
 
 .icons {
   display: flex;
   align-items: center;
-  justify-content: flex-end;
-  flex: 1;
-  height: 100%;
-}
+  justify-content: space-between;
 
-.iconLink {
-  position: relative;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  height: 64px;
-  width: 64px;
-  border-left: 1px solid $surface-2;
-  border-bottom: 3px solid transparent;
-  box-sizing: border-box;
-  transition:
-    background-color 0.3s ease,
-    border-color 0.3s ease;
+  @media (min-width: 768px) {
+    display: block;
+    align-items: center;
+    height: 100%;
 
-  img {
-    filter: $filter;
+    & > div {
+      display: flex;
+      height: 100%;
+    }
   }
-
-  &:hover {
-    background-color: $surface-1;
-  }
-}
-
-.iconLink.isActive {
-  border-bottom-color: $accent;
-  background-color: $surface-1;
 }
 
 .searchContainer {
-  flex: 1;
-  max-width: 320px;
-  margin: 0 16px 0 auto;
   position: relative;
+
+  @media (min-width: 768px) {
+    order: 0;
+    width: auto;
+    flex: 1;
+    max-width: 250px;
+    margin: 0 16px;
+  }
+
+  @media (min-width: 1024px) {
+    max-width: 350px;
+    margin: 0 32px;
+  }
 }
 
 .searchInput {
   width: 100%;
-  height: 40px;
-  padding: 0 40px 0 16px;
+  height: 36px;
+  padding: 0 36px 0 12px;
   background-color: $background;
   color: $text-main;
   border: 1px solid $surface-2;
@@ -127,6 +215,11 @@
   font-size: 14px;
   outline: none;
   transition: all 0.3s ease;
+
+  @media (min-width: 1024px) {
+    height: 40px;
+    padding: 0 40px 0 16px;
+  }
 
   &::placeholder {
     color: $text-secondary;
@@ -138,7 +231,7 @@
 
   &:focus {
     border-color: $accent;
-    box-shadow: 0 0 0 3px rgba(144, 91, 255, 0.15);
+    box-shadow: $shadow;
   }
 
   &::-webkit-search-decoration,
@@ -158,34 +251,69 @@
   transform: translateY(-50%);
   width: 16px;
   height: 16px;
-  background-image: url('/public/img/icons/search.svg');
+  background-image: url('/img/icons/search.svg');
   background-repeat: no-repeat;
   background-position: center;
   pointer-events: none;
   filter: $filter;
 }
 
-.addedProdutcs {
-  position: absolute;
-  top: 14px;
-  right: 14px;
-
+.burger {
+  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
+  width: 48px;
+  height: 48px;
+  background: none;
+  border: none;
+  border-left: 1px solid $surface-2;
+  cursor: pointer;
+  padding: 0;
+  margin-right: -16px;
+  z-index: 101;
 
-  min-width: 16px;
-  height: 16px;
-  padding: 0 4px;
-  border-radius: 10px;
+  @media (min-width: 768px) {
+    display: none;
+  }
 
-  background-color: $red;
-  border: 1px solid $background;
-  color: $text-main;
+  span,
+  span::before,
+  span::after {
+    content: '';
+    position: absolute;
+    width: 20px;
+    height: 2px;
+    background-color: $text-main;
+    transition:
+      transform 0.2s ease,
+      opacity 0.2s ease;
+  }
 
-  font-weight: 700;
-  font-size: 9px;
-  line-height: 1;
+  span {
+    left: calc(50% - 10px);
+  }
+  span::before {
+    top: -6px;
+    left: 0;
+  }
+  span::after {
+    top: 6px;
+    left: 0;
+  }
+}
 
-  box-sizing: border-box;
+.burgerActive {
+  border-left-color: transparent;
+  span {
+    background-color: transparent;
+  }
+  span::before {
+    transform: rotate(45deg);
+    top: 0;
+  }
+  span::after {
+    transform: rotate(-45deg);
+    top: 0;
+  }
 }
