+++ src/modules/shared/components/ProductCard/ProductCard.module.scss
@@ -93,9 +93,10 @@
   border: none;
   border-radius: 4px;
   cursor: pointer;
+
   transition:
     background-color 0.3s ease,
-    color 0.3s ease;
+    border-color 0.3s ease;
 
   @include buttons;
