+++ src/modules/shared/components/UI/Icons/index.ts
@@ -1,0 +1,1 @@
+export * from './Icons';
