+++ src/modules/shared/components/UI/ThemeToggler/ThemeToggler.module.scss
@@ -1,31 +1,43 @@
-// src/components/ThemeToggler/ThemeToggler.module.scss
 @import '../../../../../styles/variables';
 
 .togglerWrapper {
   display: flex;
   align-items: center;
-  justify-content: flex-end;
-  height: 64px;
-  width: 64px;
+  justify-content: center;
+  height: 56px;
+  width: 56px;
   box-sizing: border-box;
-
-  border-left: 1px solid $surface-2;
-
+  border: 1px solid $surface-2;
+  border-radius: 8px;
   transition: border-color 0.3s ease;
+
+  @media (min-width: 768px) {
+    width: 48px;
+    height: 100%;
+    border: none;
+    border-left: 1px solid $surface-2;
+    border-radius: 0;
+  }
+
+  @media (min-width: 1024px) {
+    height: 100%;
+    width: 64px;
+  }
 }
 
 .toggler {
   display: inline-flex;
   align-items: center;
   justify-content: center;
-  width: 40px;
-  height: 40px;
+  width: 36px;
+  height: 36px;
   border: 1px solid $elements;
   border-radius: 50%;
   background-color: $surface-1;
   cursor: pointer;
   transition: all 0.3s ease;
-  margin: 0 12px;
+  padding: 0;
+  margin: 0;
 
   &:hover {
     background-color: $surface-2;
@@ -33,12 +45,26 @@
     transform: scale(1.05);
   }
 
+  &:focus-visible {
+    outline: 2px solid $accent;
+    outline-offset: 2px;
+  }
+
   &:active {
     transform: scale(0.95);
+  }
+
+  @media (min-width: 1024px) {
+    width: 40px;
+    height: 40px;
   }
 }
 
 .icon {
-  font-size: 18px;
+  font-size: 16px;
   line-height: 1;
+
+  @media (min-width: 1024px) {
+    font-size: 18px;
+  }
 }
